start point to custom workload `Aerospike Workload` * add config property to disable hashing the secondary key
also make some changes to number generation
- only use longs when generating keysequences
- fix interface types for aeropike workload
- make key insert function use longs in *all* places

Introduce new number generator and use longs in measurements and
counters
   o fix increase logic; also upate the interval
   o add keychooser with growing interval

adapt copyright statements

+++ core/src/main/java/site/ycsb/measurements/OneMeasurementTimeSeries.java
@@ -60,7 +60,7 @@ public class OneMeasurementTimeSeries extends OneMeasurement {
   private long totallatency = 0;
 
   //keep a windowed version of these stats for printing status
-  private int windowoperations = 0;
+  private long windowoperations = 0;
   private long windowtotallatency = 0;
 
   private int min = -1;
